+++ tests/test_samplers/test_nested_sampler/test_general_config.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 from unittest.mock import patch, MagicMock
 
-from nessai.nestedsampler import NestedSampler
+from nessai.samplers.nestedsampler import NestedSampler
 
 
 def test_init(sampler):
@@ -36,50 +36,19 @@ def test_init(sampler):
     model.configure_pool.assert_called_once_with(pool=pool, n_pool=n_pool)
 
 
-@patch("numpy.random.seed")
-@patch("torch.manual_seed")
-def test_set_random_seed(mock1, mock2, sampler):
-    """Test the correct functions are called when setting the random seed"""
-    NestedSampler.setup_random_seed(sampler, 150914)
-    mock1.assert_called_once_with(150914)
-    mock2.assert_called_once_with(seed=150914)
-
-
-@patch("numpy.random.seed")
-@patch("torch.manual_seed")
-def test_no_random_seed(mock1, mock2, sampler):
-    """Assert no seed is set if seed=None"""
-    NestedSampler.setup_random_seed(sampler, None)
-    mock1.assert_not_called()
-    mock2.assert_not_called()
-
-
-def test_setup_output(sampler, tmpdir):
+@pytest.mark.parametrize("plot", [False, True])
+def test_setup_output(sampler, tmpdir, plot):
     """Test setting up the output directories"""
-    p = tmpdir.mkdir("outputs")
-    sampler.plot = False
-    path = os.path.join(p, "tests")
-    resume_file = NestedSampler.setup_output(sampler, path)
-    assert os.path.exists(path)
-    assert resume_file == os.path.join(path, "nested_sampler_resume.pkl")
-
-
-def test_setup_output_w_plotting(sampler, tmpdir):
-    """Test setting up the output directories with plot=True"""
-    p = tmpdir.mkdir("outputs")
-    sampler.plot = True
-    path = os.path.join(p, "tests")
-    NestedSampler.setup_output(sampler, path)
-    assert os.path.exists(os.path.join(path, "diagnostics"))
-
-
-def test_setup_output_w_resume(sampler, tmpdir):
-    """Test output configuration with a specified resume file"""
-    p = tmpdir.mkdir("outputs")
-    sampler.plot = False
-    path = os.path.join(p, "tests")
-    resume_file = NestedSampler.setup_output(sampler, path, "resume.pkl")
-    assert resume_file == os.path.join(path, "resume.pkl")
+    path = str(tmpdir.mkdir("test"))
+    sampler.plot = plot
+    rf = "test.pkl"
+    with patch(
+        "nessai.samplers.base.BaseNestedSampler.configure_output"
+    ) as mock:
+        NestedSampler.configure_output(sampler, path, rf)
+    mock.assert_called_once_with(path, rf)
+    if plot:
+        assert os.path.exists(os.path.join(path, "diagnostics"))
 
 
 def test_configure_max_iteration(sampler):